Refactor apts/skyfield_searches/visibility.py into modular package
- Split 507-line 'God Object' into twilight, rise_set, and culmination modules.
- Maintain public API via __init__.py for backward compatibility.
- Update internal relative imports.
- Pass existing skyfield search unit tests.
